+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Shared zeroconf instances can now be used
+- More units for multi level sensors
 
 ### Fixed